Enhanced creating course route and students can be added to a course

+++ src/main/java/com/hva/aad/SmallEducator/dao/StudentDao.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentDao extends CrudRepository<StudentModel, Long> {
+public interface StudentDao extends CrudRepository<StudentModel, Integer> {
 }
